appflask/requirements.txt
wheel
aniso8601
certifi
chardet
Click
Flask
Flask-jwt-extended
Flask-RESTful
Flask-SQLAlchemy
idna
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
PyJWT
python-dotenv
pytz
requests
six
urllib3
uWSGI
Werkzeug
jsonschema
flask-bcrypt

